requirements.txt
# Core ML Stack
torch>=2.1.0
transformers>=4.36.0
trl>=0.12.0
peft>=0.7.0
accelerate>=0.25.0
datasets>=2.15.0

# Inference
llama-cpp-python>=0.2.0

# Memory (RAG)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data Pipeline
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization & UI
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
